requirements.txt
streamlit
inference
Pillow
